Remove User Preferences
Warning on leaving an edit page was removed, this functionality
is handled by visualeditor.
Removed autosaving artifacts and more obsolete functionality.

Left only code behind which is needed for concurrent edit warnings.

Needs more refactoring, as functionality of edit warnings is now the
only function of this extension, and this warning is still not shown
in visual editor.

ERM #10122
[3.1]

Change-Id: Iaac45c371de110cdeb004e72583425526864c357